Chrome outlining
Unlike Safari, where the scrollbar is not included in the screenwidth,
Chrome does include the side scrollbar to the screenwidth. This causes
a difference between the widthwrap (fixed width so padding is changed
when scrollbar is added) or an element (like the header) that does not
have a fixed width but a fixed padding (the change is made in the
center of the element).
Fixed this by using the widthwrap everywhere; this also made the CSS
shorter.

In general: be consistent with either using same side-paddings
everywhere and no fixed width. Or use fixed-width everywhere and no
side-paddings. Two systems together make inconsistent outlining in
Chrome.